Create Uniform_Cost_Search.py
It is the algorithm which gives you the minimal path cost from start node to goal node by exploring the different paths. It works for large for large graphs as a alternative to the Dijkstra's algorithm. Mainly its application is found in AI search and problem solving methods and it is also incorporated into the A* search along with the heuristics.